refactored and added tests for psc type handler and router
