+++ test/routers/pscTypeRouter.int.ts
@@ -1,3 +1,4 @@
+import * as cheerio from "cheerio";
 import middlewareMocks from "../mocks/allMiddleware.mock";
 import request from "supertest";
 import { PrefixedUrls } from "../../src/constants";
@@ -20,7 +21,7 @@ beforeEach(() => {
     jest.clearAllMocks();
 });
 
-describe("psc type tests", () => {
+describe("psc type router tests", () => {
 
     beforeEach(() => {
         middlewareMocks.mockSessionMiddleware.mockClear();
@@ -30,24 +31,43 @@ describe("psc type tests", () => {
         expect(middlewareMocks.mockSessionMiddleware).toHaveBeenCalledTimes(1);
     });
 
-    it("Should render the PSC Type page with a successful status code", async () => {
-        const resp = await request(app).get(PrefixedUrls.PSC_TYPE);
-        expect(resp.status).toBe(200);
+    describe("Test router.get", () => {
+
+        it("Should render the PSC Type page with a successful status code", async () => {
+            const resp = await request(app).get(PrefixedUrls.PSC_TYPE);
+            expect(resp.status).toBe(200);
+        });
+
+        it.each([[undefined, undefined], ["individual", "pscType=individual"], ["rle", "pscType=rle"]])("Should render the Psc Type page with a success status code and %s radio button checked", async (expectedSelection, expectedQuery) => {
+            const queryParams = new URLSearchParams(expectedQuery);
+            const uriWithQuery = `${PrefixedUrls.PSC_TYPE}?${queryParams}`;
+            const uri = getUrlWithTransactionIdAndSubmissionId(uriWithQuery, TRANSACTION_ID, PSC_VERIFICATION_ID);
+
+            const resp = await request(app).get(uri);
+
+            const $ = cheerio.load(resp.text);
+
+            expect(resp.status).toBe(HttpStatusCode.Ok);
+            expect($("input[name=pscType]:checked").val()).toBe(expectedSelection);
+        });
     });
 
-    it.each([["individual", PrefixedUrls.INDIVIDUAL_PSC_LIST], ["rle", PrefixedUrls.RLE_LIST], ["default", PrefixedUrls.INDIVIDUAL_PSC_LIST]])(
-        "Should redirect to %s list page if selected",
-        async (selectedType, expectedPage) => {
+    describe("Test router.post", () => {
 
-            const expectedRedirectUrl = `${expectedPage.replace(":transactionId", TRANSACTION_ID).replace(":submissionId", PSC_VERIFICATION_ID)}?companyNumber=${COMPANY_NUMBER}&lang=en&pscType=${selectedType}`;
+        it.each([["individual", PrefixedUrls.INDIVIDUAL_PSC_LIST], ["rle", PrefixedUrls.RLE_LIST], ["default", PrefixedUrls.INDIVIDUAL_PSC_LIST]])(
+            "Should redirect the post request to %s list page if selected",
+            async (selectedType, expectedPage) => {
 
-            await request(app)
-                .post(getUrlWithTransactionIdAndSubmissionId(PrefixedUrls.PSC_TYPE, TRANSACTION_ID, PSC_VERIFICATION_ID))
-                .send({ pscType: selectedType })
-                .set({ "Content-Type": "application/json" })
-                .query({ companyNumber: COMPANY_NUMBER, lang: "en", pscType: selectedType })
-                .expect(HttpStatusCode.Found)
-                .expect("Location", expectedRedirectUrl);
-        }
-    );
+                const expectedRedirectUrl = `${expectedPage.replace(":transactionId", TRANSACTION_ID).replace(":submissionId", PSC_VERIFICATION_ID)}?companyNumber=${COMPANY_NUMBER}&lang=en&pscType=${selectedType}`;
+
+                await request(app)
+                    .post(getUrlWithTransactionIdAndSubmissionId(PrefixedUrls.PSC_TYPE, TRANSACTION_ID, PSC_VERIFICATION_ID))
+                    .send({ pscType: selectedType })
+                    .set({ "Content-Type": "application/json" })
+                    .query({ companyNumber: COMPANY_NUMBER, lang: "en", pscType: selectedType })
+                    .expect(HttpStatusCode.Found)
+                    .expect("Location", expectedRedirectUrl);
+            }
+        );
+    });
 });
